add local packages as dependencies for werift-webrtc
the multicast-dns package was added to werift-ice that was not being installed with
werift-webrtc. This is because we import the werift-ice code with
relative imports, i.e.
import { Connection } from "../../../ice/src";

This was making it impossible to use werift-webrtc without manually adding
multicast-dns as a dependency of whatever project is using werift-webrtc

Adding werift-ice (and the other local packages) as a dependency of
werift-webrtc fixes this issue and any future dependencies added
won't cause problems